+++ src/code/functions/meta.py
@@ -1,21 +1,18 @@
-import sys
-from core.interactions import Response
 from core.tokenizer import tokenizer
 from connectors.selector import get_db_connection
 
 
-def get_info(request, data_type="meta"):
+def get_info(request, response, data_type="meta"):
+    resp = response
 
-    resp = Response()
-    
     connection = get_db_connection(request.json_data.get("path", {}).get("connection"), database=request.json_data.get("path", {}).get("database"))
     if connection is None:
         resp.output({ "ok": False, "error": "Invalid connection specified." })
-        sys.exit()
+        return
 
     if not connection.open():
         resp.output({ "ok": False, "error": "Unable to connect to server.  Please check username/password." })
-        sys.exit()
+        return
     
     in_path = request.json_data.get("path", {})
     for item in in_path:
@@ -27,18 +24,18 @@ def get_info(request, data_type="meta"):
 
         tokenizer.update()
         resp.output({ "ok": True, "meta": meta, "items": data })
-        sys.exit()
+        return
 
     if data_type == "ddl":
         meta, statement = connection.ddl(str(request.json_data.get("type"))[0:100], str(request.json_data.get("target"))[0:255], in_path)
 
         tokenizer.update()
         resp.output({ "ok": True, "meta": meta, "ddl": statement })
-        sys.exit()
+        return
 
     if data_type == "details":
         details = connection.details(str(request.json_data.get("type"))[0:100], str(request.json_data.get("target"))[0:255], in_path)
 
         tokenizer.update()
         resp.output({ "ok": True, "properties": details })
-        sys.exit()
+        return