+++ src/code/index.py
@@ -17,12 +17,15 @@ logging.getLogger("botocore").setLevel(logging.INFO)
 logging.getLogger("urllib3").setLevel(logging.INFO)
 
 request = Request(**os.environ)
+resp = Response()
+
 if request.headers.get("CONTENT_LENGTH", 0) > 0:
     request.set_data(sys.stdin.read(request.headers.get("CONTENT_LENGTH")))
 else:
     logging.error(f"Invalid request from {request.host}")
-    resp = Response()
     resp.output({ "ok": False })
+    resp.send()
     sys.exit()
 
-process_request(request)
+process_request(request, resp)
+resp.send()